+++ Sources/Differentia+PrettyPrintable.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+
+
+extension Differentia: PrettyPrintable {
+    var prettyLines: [PrettyLine] {
+        return self.units
+            .flatMap { unit -> [PrettyLine] in unit.prettyLines }
+    }
+}
